requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
jinja2>=3.1.0
schedule>=1.2.0
python-dateutil>=2.8.0
colorama>=0.4.6

# X API (optional — only needed if using official API)
# tweepy>=4.14.0

# API Server for Chrome Extension
Flask>=3.0.0
Flask-Cors>=4.0.0
